Finished Upcoming elections part: Added Mailing address, Added URL loading, Added HideInfoIfNull Bind URLs to TextViews Fixed ProgressBar

+++ app/src/main/java/com/example/android/politicalpreparedness/network/CivicsApiService.kt
@@ -44,7 +44,7 @@ interface CivicsApiService {
     @GET("voterinfo")
     suspend fun voterInfoQuery(
         @Query("address") address: String,
-        @Query("electionId") electionId: Long
+        @Query("electionId") electionId: Int
     ) : VoterInfoResponse
 
     //Added representatives API Call
